Prod_CV_NLP_API/flask/requirements.txt
Flask==1.0.2
PassportEye==1.4.0
pillow==6.2.0
pytesseract==0.2.6
opencv-python==4.1.0.25
pdf2image==1.14.0
flask-cors==3.0.7
torch==1.7.1
transformers==4.0.1
python-docx==0.8.10
regex==2020.11.13
pdftotext==2.1.5
ipywidgets
setuptools
joblib
numpy
easyocr
pyyaml
watchdog

